requirements.txt
amqp==2.6.1
async-timeout==3.0.1
billiard==3.6.3.0
celery==4.3.0
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-anymail==6.1.0
django-celery-beat==1.4.0
django-cors-headers==2.5.2
django-countries==7.0
django-filter==2.1.0
django-redis==4.12.1
django-rest-swagger==2.2.0
django-timezone-field==4.1.1
djangorestframework==3.9.2
djangorestframework-filters==0.11.1
hiredis==1.1.0
idna==2.8
importlib-metadata==3.7.0
itypes==1.2.0
Jinja2==2.11.3
kaleido==0.2.0
kombu==4.6.11
MarkupSafe==1.1.1
openapi-codec==1.3.2
plotly==4.14.3
psutil==5.8.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-http-client==3.3.2
pytz==2021.1
redis==3.2.1
requests==2.21.0
retrying==1.3.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.24.3
vine==1.3.0
zipp==3.4.0
